+++ src/OpenSSL/crypto.py
@@ -1010,6 +1010,12 @@ class X509Req:
         :param int version: The version number.
         :return: ``None``
         """
+        if not isinstance(version, int):
+            raise TypeError("version must be an int")
+        if version != 0:
+            raise ValueError(
+                "Invalid version. The only valid version for X509Req is 0."
+            )
         set_result = _lib.X509_REQ_set_version(self._req, version)
         _openssl_assert(set_result == 1)
 
@@ -1639,7 +1645,6 @@ class X509StoreFlags:
     POLICY_CHECK: int = _lib.X509_V_FLAG_POLICY_CHECK
     EXPLICIT_POLICY: int = _lib.X509_V_FLAG_EXPLICIT_POLICY
     INHIBIT_MAP: int = _lib.X509_V_FLAG_INHIBIT_MAP
-    NOTIFY_POLICY: int = _lib.X509_V_FLAG_NOTIFY_POLICY
     CHECK_SS_SIGNATURE: int = _lib.X509_V_FLAG_CHECK_SS_SIGNATURE
     PARTIAL_CHAIN: int = _lib.X509_V_FLAG_PARTIAL_CHAIN